Added loading message
Caching of work items was also removed since it messed with tasks that were cached before having their states changed,
which resulted in wrong information being relayed. This wasn't that common, since dailies (usually) only occur once,
but it was annoying either way. Now at least a message is there to show the bot is alive and working on the request.